+++ lib/src/code_generator/struc.dart
@@ -44,6 +44,9 @@ class Struc extends NoLookUpBinding {
 
   bool get isOpaque => members.isEmpty;
 
+  /// Value for `@Packed(X)` annotation. Can be null(no packing), 1, 2, 4, 8, 16.
+  int? pack;
+
   /// Marker for checking if the dependencies are parsed.
   bool parsedDependencies = false;
 
@@ -52,6 +55,7 @@ class Struc extends NoLookUpBinding {
     String? originalName,
     required String name,
     this.isInComplete = false,
+    this.pack,
     String? dartDoc,
     List<Member>? members,
   })  : members = members ?? [],
@@ -105,6 +109,10 @@ class Struc extends NoLookUpBinding {
     /// to have the same name as the class.
     final localUniqueNamer = UniqueNamer({enclosingClassName});
 
+    /// Write @Packed(X) annotation if struct is packed.
+    if (pack != null) {
+      s.write('@${w.ffiLibraryPrefix}.Packed($pack)\n');
+    }
     // Write class declaration.
     s.write(
         'class $enclosingClassName extends ${w.ffiLibraryPrefix}.${isOpaque ? 'Opaque' : 'Struct'}{\n');