+++ lib/src/config_provider/config_types.dart
@@ -31,6 +31,35 @@ enum CommentLength { none, brief, full }
 
 enum StructDependencies { full, opaque }
 
+/// Holds config for how Structs Packing will be overriden.
+class StructPackingOverride {
+  final Map<RegExp, int?> _matcherMap;
+
+  StructPackingOverride({Map<RegExp, int?>? matcherMap})
+      : _matcherMap = matcherMap ?? {};
+
+  /// Returns true if the user has overriden the pack value.
+  bool isOverriden(String name) {
+    for (final key in _matcherMap.keys) {
+      if (quiver.matchesFull(key, name)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  /// Returns pack value for [name]. Ensure that value [isOverriden] before
+  /// using the returned value.
+  int? getOverridenPackValue(String name) {
+    for (final opv in _matcherMap.entries) {
+      if (quiver.matchesFull(opv.key, name)) {
+        return opv.value;
+      }
+    }
+    return null;
+  }
+}
+
 /// Represents a single specification in configurations.
 ///
 /// [E] is the return type of the extractedResult.