+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.5.0-beta.1
+- Added support for `Packed` structs. Packed annotations are generated
+automatically but can be overriden using `structs -> pack` config.
+- Updated sdk constraints to `>=2.13.0-211.6.beta`.
+
 # 2.4.2
 - Fix issues due to declarations having duplicate names.
 - Fix name conflict of declaration with ffi library prefix.